working snmpv3

+++ modules/snmp/gateway/src/main/kotlin/com/mussonindustrial/embr/snmp/configuration/settings/SnmpDeviceSettings.kt
@@ -3,7 +3,7 @@ package com.mussonindustrial.embr.snmp.configuration.settings
 interface SnmpDeviceSettings {
     val hostname: String
     val port: Int
-    val timeout: Long
+    val timeout: Int
     val healthcheckFrequency: Int?
     val healthcheckOid: String?
 }
