+++ modules/snmp/gateway/src/main/kotlin/com/mussonindustrial/embr/snmp/configuration/types/SnmpV1DeviceType.kt
@@ -55,13 +55,13 @@ object SnmpV1DeviceType :
         override val readTarget =
             CommunityTarget(address, OctetString(snmpSettings.communityRead)).apply {
                 version = SnmpConstants.version1
-                timeout = snmpSettings.timeout
+                timeout = snmpSettings.timeout.toLong()
             }
         override val writeTarget =
             snmpSettings.communityWrite?.let {
                 CommunityTarget(address, OctetString(snmpSettings.communityWrite)).apply {
                     version = SnmpConstants.version1
-                    timeout = snmpSettings.timeout
+                    timeout = snmpSettings.timeout.toLong()
                 }
             }
